- Move test files to be with their components - Minor jest config changes - use npx react-native in package.json scripts
