+++ examples/react-native/v12/TestApp/App.tsx
@@ -5,12 +5,12 @@ import {NavigationContainer} from '@react-navigation/native';
 import {createDrawerNavigator} from '@react-navigation/drawer';
 
 import {HomeScreen} from './src/screens/HomeScreen';
-import {Geospatial} from './src/components/data-types/Geospatial';
+import {Geospatial} from './src/components/data-types/geospatial/Geospatial';
 
-import {FtsQuery} from './src/components/FtsQuery';
+import {FtsQuery} from './src/components/fts-query/FtsQuery';
 
-import {Logger} from './src/components/Logger';
-import {ObjectModels} from './src/components/ObjectModels';
+import {Logger} from './src/components/logger/Logger';
+import {ObjectModels} from './src/components/object-models/ObjectModels';
 import {SubscriptionScreen} from './src/screens/SubscriptionScreen';
 
 import {RootStackParamList} from './src/navigation/types';
@@ -35,7 +35,7 @@ function App(): JSX.Element {
       <Drawer.Navigator initialRouteName="Home">
         <Drawer.Screen name="Home" component={HomeScreen} />
         <Drawer.Screen name="Geospatial" component={Geospatial} />
-        <Drawer.Screen name="Full-Text Search" component={FtsQuery} />
+        <Drawer.Screen name="FullTextSearch" component={FtsQuery} />
         <Drawer.Screen name="Logger" component={Logger} />
         <Drawer.Screen name="ObjectModels" component={ObjectModels} />
         <Drawer.Screen name="Subscriptions" component={SubscriptionScreen} />
