--- examples/react-native/v12/TestApp/src/__tests__/quickstart/readme.md
@@ -1,1 +1,0 @@
-Placeholder file to keep directory structure.
