+++ examples/react-native/v12/TestApp/src/components/data-types/geospatial/Geospatial.test.tsx
@@ -1,6 +1,6 @@
 import 'react-native';
 import React from 'react';
-import {Geospatial} from '../../components/data-types/Geospatial';
+import {Geospatial} from './Geospatial';
 import {render, screen} from '@testing-library/react-native';
 
 test('Geospatial queries', async () => {
