+++ examples/react-native/v12/TestApp/src/components/data-types/geospatial/Geospatial.tsx
