+++ examples/react-native/v12/TestApp/src/components/fts-query/FtsQuery.tsx
