+++ examples/react-native/v12/TestApp/src/components/logger/Logger.tsx
