+++ examples/react-native/v12/TestApp/src/components/object-models/ObjectModels.test.tsx
@@ -1,6 +1,6 @@
 import 'react-native';
 import React from 'react';
-import {ObjectModels} from '../../components/ObjectModels';
+import {ObjectModels} from './ObjectModels';
 import {render, screen, within} from '@testing-library/react-native';
 
 test('Object models', async () => {
