+++ examples/react-native/v12/TestApp/src/components/subscribe-api/BasicSubscription.tsx
