+++ source/sdk/react-native/sync-data/flexible-sync.txt
@@ -177,7 +177,7 @@ to manually manage a set of subscriptions to specific queries on queryable
 fields.
 
 You can use the :realm-react-sdk:`Realm.subscriptions 
-<classes/Realm-1.html#subscriptions>` API to manually manage a set of
+<classes/Realm-1.html#subscriptions>` API to manage a set of
 subscriptions to specific queries on queryable fields.
 
 If you're using ``@realm/react``, you can manage realm subscriptions inside of
